requirements.txt
Flask
scikit-learn
numpy
Pillow
